+++ packages/orq-rc/src/orq_ai_sdk/models/createcontactop.py
@@ -132,7 +132,7 @@ class CreateContactResponseBody(BaseModel):
     created: Optional[datetime] = None
     r"""The date and time the resource was created"""
 
-    updated: Optional[datetime] = parse_datetime("2026-01-30T05:18:02.987Z")
+    updated: Optional[datetime] = parse_datetime("2026-01-30T11:22:44.854Z")
     r"""The date and time the resource was last updated"""
 
     @model_serializer(mode="wrap")