+++ packages/orq-rc/src/orq_ai_sdk/models/updatedatasourceop.py
@@ -104,7 +104,7 @@ class UpdateDatasourceResponseBody(BaseModel):
     r"""The number of chunks in the datasource"""
 
     id: Annotated[Optional[str], pydantic.Field(alias="_id")] = (
-        "01KG6NCWRX42828TB5MEVRRVSP"
+        "01KG7A8Q28RHDTS4EDH0JXSVMJ"
     )
     r"""The unique identifier of the data source"""