+++ packages/orq-rc/RELEASES.md
@@ -5129,3 +5129,13 @@ Based on:
 - [python v4.3.0-rc.11] packages/orq-rc
 ### Releases
 - [PyPI v4.3.0-rc.11] https://pypi.org/project/orq-ai-sdk/4.3.0-rc.11 - packages/orq-rc
+
+## 2026-01-30 11:25:36
+### Changes
+Based on:
+- OpenAPI Doc  
+- Speakeasy CLI 1.700.2 (2.801.2) https://github.com/speakeasy-api/speakeasy
+### Generated
+- [python v4.3.0-rc.12] packages/orq-rc
+### Releases
+- [PyPI v4.3.0-rc.12] https://pypi.org/project/orq-ai-sdk/4.3.0-rc.12 - packages/orq-rc